css styling on buttons, destroy all chores and children associated when destroy parent, finish all background screens

+++ db/migrate/20201116225426_create_chores.rb
@@ -4,7 +4,7 @@ class CreateChores < ActiveRecord::Migration
       t.string :name
       t.integer :parent_id
       t.integer :child_id, null: true
-      t.float :value 
+      t.integer :value 
       t.integer :tagged_complete
       
     end
